requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Model inference
llama-cpp-python==0.2.56
huggingface-hub==0.20.3

# Environment and configuration
python-dotenv==1.0.1

# Logging and utilities
pydantic==2.10.4
pydantic-settings==2.7.0

# CORS support
python-jose==3.3.0
